% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBases.R
\name{addBases}
\alias{addBases}
\title{Extract bases flanking the mutation.}
\usage{
addBases(maf, ref_genome, up = 1, down = 1)
}
\arguments{
\item{maf}{output generated by \code{read.maf}}

\item{ref_genome}{faidx indexed refrence fasta file.}

\item{up}{how many bases upstream of Start_position. Defaults to 1}

\item{down}{how many bases downstream of Start_position. Defaults to 1}
}
\value{
adds an extra column 'bases' to data slot of \code{maf}
}
\description{
Extract adjacent bases to mutation site. Defaults to one base immediate to 5' and 3' of Start_position.
Make sure chromosome names in 'Chromosome' field of maf matches to chromosome names in reference fasta file. This is useful for analyzing mutational signatures.
}
\details{
This function loads reference genome into memeory. Typical human geneome occupies a peak memory of ~3 gb while extracting bases.
}

