% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainSummary.R
\name{addPfamDomains}
\alias{addPfamDomains}
\title{pfam domain annotation.}
\usage{
addPfamDomains(maf = NULL, AACol = NULL, summarizeBy = "AAPos")
}
\arguments{
\item{maf}{an MAF object generated by \code{\link{read.maf}}}

\item{AACol}{manually specify column name for amino acid changes. Default looks for field 'AAChange'}

\item{summarizeBy}{Summarize domains by amino acid position or conversions. Can be "AAPos" or "AAChange"}
}
\value{
Summarized data table for each amino acid conversion with pfam domain information.
}
\description{
Summarizes amino acid positions and annotates them with pfam domain information.
}

