% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addReadCounts.R
\name{addReadCounts}
\alias{addReadCounts}
\title{add read counts to MAF file from bam files.}
\source{
bam-readcount: https://github.com/genome/bam-readcount
}
\usage{
addReadCounts(maf, bam_df, BaseQuality = 10, MapQuality = 10,
  ref_genome = NULL, skipSamples = NULL, samples = NULL, prefix = NULL,
  add = TRUE)
}
\arguments{
\item{maf}{an MAF object generated by \code{\link{read.maf}}}

\item{bam_df}{two column dataframe. First column must contain Tumor_Sample_Barcode and second column with path to corresponding bam file.}

\item{BaseQuality}{minimum base quality.}

\item{MapQuality}{minimum mapping quality.}

\item{ref_genome}{path to reference genome. required.}

\item{skipSamples}{Tumor_Sample_Barcodes to skip from bam_df.}

\item{samples}{add read counts to only these samples.}

\item{prefix}{Prefix to add or remove from contig names in MAF file.}

\item{add}{If prefix is used, default is to add prefix to contig names in MAF file. If false prefix will be removed from contig names.}
}
\value{
maf file with "t_depth", "t_ref_count", "t_alt_count" and "t_vaf" columns added.
}
\description{
add read counts to MAF file from bam files.
}
\details{
This is a wrapper script which runs bam-readcount programme via R \code{system()} command to get readcounts.
Assumes bam-readcount is installed and under path.
Make sure chromosome names in 'Chromosome' field of maf matches to chromosome names in reference fasta file.
Currently only supports for SNPs. Indels will be populated with NA.
}

