% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerhotspots.R
\name{cancerhotspots}
\alias{cancerhotspots}
\title{extract nucleotide counts/variant allele frequencies of targeted (somatic) variants from the BAM file.}
\usage{
cancerhotspots(
  bam = NULL,
  refbuild = "GRCh37",
  mapq = 10,
  sam_flag = 1024,
  vaf = 0.05,
  t_depth = 30,
  t_alt_count = 8,
  op = NULL,
  fa = NULL,
  browse = FALSE
)
}
\arguments{
\item{bam}{Input bam file. Required.}

\item{mapq}{Map quality. Default 10}

\item{sam_flag}{SAM FLAG to filter reads. Default 1024}

\item{vaf}{VAF threshold. Default 0.05 [Variant filter]}

\item{t_depth}{Depth of coverage threshold. Default 30 [Variant filter]}

\item{t_alt_count}{Min. number of reads supporting tumor allele . Default 8 [Variant filter]}

\item{op}{Output file basename. Default parses from BAM file}

\item{fa}{Indexed fasta file. If provided, extracts and adds reference base to the output tsv.}

\item{browse}{If TRUE opens the html file in browser}

\item{bed}{Input BED file. Required. Should be one based coordinate}
}
\description{
extract nucleotide counts/variant allele frequencies of targeted (somatic) variants from the BAM file.
}
