% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerhsAggr.R
\name{cancerhotspotsAggr}
\alias{cancerhotspotsAggr}
\title{Aggregate cancerhotspots reports}
\usage{
cancerhotspotsAggr(
  tsvs = NULL,
  minVaf = 0.02,
  minDepth = 15,
  sampleNames = NULL,
  maf = TRUE,
  ...
)
}
\arguments{
\item{tsvs}{TSV files generated by \code{\link{cancerhotspots}}}

\item{minVaf}{Min. VAF threshold. Default 0.02}

\item{minDepth}{Min. depth of coverage. Default 15}

\item{sampleNames}{samples for each tsv file. Default NULL. Parses from file names.}

\item{maf}{Return as an MAF object. Default TRUE.}

\item{...}{Additional argumnets passed to \code{\link{read.maf}} if `maf` is TRUE.}
}
\value{
\code{\link{MAF}} object
}
\description{
Takes tsv files generated by \code{\link{cancerhotspots}} and aggregates them into an MAF for downstream analysis
}
\seealso{
\code{\link{cancerhotspots}}
}
