% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coGisticChromPlotV.R
\name{coGisticChromPlotV}
\alias{coGisticChromPlotV}
\title{Co-plot version of gisticChromPlot()}
\usage{
coGisticChromPlotV(
  gistic1 = NULL,
  gistic2 = NULL,
  g1Name = "",
  g2Name = "",
  type = "Amp",
  markBands = TRUE,
  labelGenes = TRUE,
  y_lims = NULL,
  maf1 = NULL,
  maf2 = NULL,
  mutGenes = NULL,
  mutGenes1 = NULL,
  mutGenes2 = NULL,
  fdrCutOff = 0.05,
  symmetric = TRUE,
  color = NULL,
  ref.build = "hg19",
  cytobandOffset = "auto",
  txtSize = 0.8,
  cytobandTxtSize = 1,
  mutGenesTxtSize = 0.6,
  rugTickSize = 0.1
)
}
\arguments{
\item{gistic1}{data will be plotted on the left side, yload_gistic() or readGistic() returned
object}

\item{gistic2}{data will be plotted on the right side, yload_gistic() or readGistic() returned
object}

\item{g1Name}{the title of the left side}

\item{g2Name}{the title of the right side}

\item{type}{default 'Amp', c('Amp',"Del"), choose one to plot}

\item{markBands}{default TRUE, integer of length 1 or 2 or TRUE, mark}

\item{labelGenes}{if you want to label some genes you are interested along the chromosome, set
TRUE}

\item{y_lims}{default NULL, auto determine the G-score ranges (x axis lims) of the plot. You can
set it to c(-6,6), the left plot values must be negative and right positve, the abs values of
the y_lims will be annotated as the x-axis tick labels. If you set y_lims to not NULL values,
then the `symmetric` will be auto set to FALSE}

\item{maf1, maf2}{if labelGenes==TRUE, you need to provide read.maf() object, the genes mutation
info collected from the maf1 is shown on the left side, while maf2 on the right side. the genes
selected are controled by the mutGenes or mutGenes1 or mutGenes1 parameter, see following.}

\item{mutGenes, mutGenes1, mutGenes2}{default NULL, could be NULL, number, or character vector of
gene symbols which match the corresponding  MAF object's Hugo_Symbol column values. mutGenes
controls both sides of the annotation, mutGenes1 controls only left side and corresponding data
is extracted from to maf1, and mutGenes2 controls only right side annotation and corresponding
to maf2. If `NULL`, extract the top 50 mutated genes from maf1 and maf2 seperatedly then
annotate them on the left side (maf1 genes) and right side (maf2 genes). if integer, say N,
only top N genes will be extracted seperately from maf1 and maf2. These two condition leads to
different genes annotated on both sides. If character vector, then the genes have mutated in
maf1 and maf2 will be annotated on both side of the figure which mean the two sides have the
same list of genes. if mutGenes is not NULL and both mutGenes1 and mutGenes1 are NULL, then the
auto set mutGenes1 = mutGenes2 = mutGenes.}

\item{fdrCutOff}{default 0.05,only items with FDR < fdrCutOff will be colored as Amp or Del (
colored 'Red' or 'Blue'), others will be seen as non-significant events (colored gray)}

\item{symmetric}{default TRUE, If False, when the gistic1 and gistic2 have different max values
of G-scores, the Chrom (0 point of x axis) will not be in the center of the whole plot, if you
set symmetric==TRUE, then the one with smaller max(G-score) will be stretched larger to make
the 0 of the x axis in the middle which eventually make the plot more symmetric.}

\item{color}{NULL or a named vector. the color of the G-score lines, default NULL which will set
the color c(Amp = "red", Del = "blue", neutral = 'gray70')}

\item{ref.build}{default "hg19", c('hg18','hg19','hg38') supported at current.(same as maftools
other functions)}

\item{cytobandOffset}{default 'auto', the width of the chromosome rects(Y axis at 0 point of X
axis). by default will be 1.5% of the width of x axis.}

\item{txtSize}{the zoom value of most of the texts}

\item{cytobandTxtSize}{textsize of the cytoband annotation}

\item{mutGenesTxtSize}{textsize of the mutGenes annotation}

\item{rugTickSize}{the rug line width of the cytoband annotation}
}
\description{
Use two GISTIC object or/and two MAF obejct to view a vertical arranged version of
  Gistic Chromosome plot results on the Amp or Del G-scores.
}
\examples{
\dontrun{
all.lesions <- system.file("extdata", "all_lesions.conf_99.txt", package = "maftools")
amp.genes <- system.file("extdata", "amp_genes.conf_99.txt", package = "maftools")
del.genes <- system.file("extdata", "del_genes.conf_99.txt", package = "maftools")
scores.gis <- system.file("extdata", "scores.gistic", package = "maftools")
laml.gistic = readGistic(gisticAllLesionsFile = all.lesions, gisticAmpGenesFile = amp.genes, gisticDelGenesFile = #' del.genes, gisticScoresFile = scores.gis, isTCGA = TRUE)

# or you can use new helper function to quickly read gistic results into new object:
gistic_res_folder = system.file("extdata",package = "maftools")
laml.gistic2 = yload_gistic(gistic_res_folder)
# NOTE:
# to quickly show the plots, I have to use the data inside the maftools, so the left and right part of the plot is mirrored, but
# in practise you will send different GISTIC and MAF objects to `gistic1`,`gistc2`, `maf1`,`maf2`, the function will also work well.

laml.maf = system.file('extdata', 'tcga_laml.maf.gz', package = 'maftools')
laml.clin = system.file('extdata', 'tcga_laml_annot.tsv', package = 'maftools')
laml = read.maf(maf = laml.maf, clinicalData = laml.clin)
laml2 = laml

# --- plot ---
gisticChromPlot2v(gistic1 = laml.gistic, gistic2 = laml.gistic2, type='Del',
                   symmetric = TRUE, g1Name = 'TCGA1',
                   g2Name = 'TCGA2', maf1 = laml, maf2 = laml2, mutGenes = 30)
}
}
