% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_methods.R
\docType{methods}
\name{getAnnotations}
\alias{getAnnotations}
\alias{getAnnotations,MAF-method}
\alias{getAnnotations}
\title{extract annotations from MAF object}
\usage{
getAnnotations(x)

\S4method{getAnnotations}{MAF}(x)
}
\arguments{
\item{x}{An object of class MAF}
}
\value{
annotations associated with samples in MAF
}
\description{
extract annotations from MAF object
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf, removeSilent = TRUE, useAll = FALSE)
getAnnotations(x = laml)
}
