% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_methods.R
\docType{methods}
\name{getClinicalData}
\alias{getClinicalData}
\alias{getClinicalData,MAF-method}
\title{extract annotations from MAF object}
\usage{
getClinicalData(x)

\S4method{getClinicalData}{MAF}(x)
}
\arguments{
\item{x}{An object of class MAF}
}
\value{
annotations associated with samples in MAF
}
\description{
extract annotations from MAF object
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
getClinicalData(x = laml)
}
