% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{getCounts}
\alias{getCounts}
\title{accessory function to \code{addReadCounts()}.}
\usage{
getCounts(var, bam, MapQuality, BaseQuality, ref_genome)
}
\description{
This is an accessory function. Use \code{addReadCounts()}
}

