% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_methods.R
\docType{methods}
\name{getFields}
\alias{getFields}
\alias{getFields,MAF-method}
\alias{getFields}
\title{extract available fields from MAF object}
\usage{
getFields(x)

\S4method{getFields}{MAF}(x)
}
\arguments{
\item{x}{An object of class MAF}
}
\value{
Field names in MAF file
}
\description{
extract available fields from MAF object
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf, removeSilent = TRUE, useAll = FALSE)
getFields(x = laml)
}
