% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_methods.R
\docType{methods}
\name{getSampleSummary}
\alias{getSampleSummary}
\alias{getSampleSummary,MAF-method}
\title{extract sample summary from MAF object}
\usage{
getSampleSummary(x)

\S4method{getSampleSummary}{MAF}(x)
}
\arguments{
\item{x}{An object of class MAF}
}
\value{
sample summary table
}
\description{
extract sample summary from MAF object
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf, removeSilent = TRUE, useAll = FALSE)
getSampleSummary(x = laml)
}

