% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTumHetero.R
\name{inferHeterogeneity}
\alias{inferHeterogeneity}
\title{Clusters variants based on Variant Allele Frequencies (VAF).}
\usage{
inferHeterogeneity(maf, tsb = NULL, top = 5, vafCol = NULL, plot = TRUE)
}
\arguments{
\item{maf}{an MAF object generated by \code{\link{read.maf}}}

\item{tsb}{specify Tumor_Sample_Barcode for which clustering has to be done.}

\item{top}{if \code{tsb} is NULL, uses top n number of most mutated samples. Defaults to 5.}

\item{vafCol}{manually specify column name for vafs. Default looks for column 't_vaf'}

\item{plot}{logical plots density curve and clusters.}
}
\value{
list of clustering tables.
}
\description{
takes output generated by read.maf and clusters variants to infer tumor heterogenity (multiple clones).
}
\details{
This function clusters variants in a particlar tumor sample by running \code{densityMclust} on vaf's to estimate univariate density and clusters.
}
\examples{
inferHeterogeneity(maf = laml, tsb = 'TCGA.AB.2972', vafCol = 'i_TumorVAF_WU')
}
\references{
Chris Fraley and Adrian E. Raftery (2002) Model-based Clustering, Discriminant Analysis and Density Estimation Journal of the American
Statistical Association 97:611-631
}

