% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinMap.R
\name{lollipopPlot}
\alias{lollipopPlot}
\title{Draws lollipop plot of amino acid changes on to Protein structure.}
\usage{
lollipopPlot(maf, gene = NULL, refSeqID = NULL, proteinID = NULL,
  label = F, AACol = NULL)
}
\arguments{
\item{maf}{an MAF object generated by \code{read.maf}}

\item{gene}{HGNC symbol for which protein structure to be drawn.}

\item{refSeqID}{RefSeq transcript identifier for \code{gene} if known.}

\item{proteinID}{RefSeq protein identifier for \code{gene} if known.}

\item{label}{if \code{TRUE} lables dots with amino acid conversions using \code{ggrepel} package.}

\item{AACol}{manually specify column name for amino acid changes. Default looks for field 'AAChange'}
}
\value{
ggplot object of the plot, which can be futher modified.
}
\description{
Draws lollipop plot of amino acid changes.
}
\details{
This function by default looks for field "AAChange" in maf file. One can also manually specify field name containing amino acid changes.
}

