% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafObjects.R
\name{maf2mae}
\alias{maf2mae}
\title{Convert MAF to MultiAssayExperiment object}
\usage{
maf2mae(m = NULL)
}
\arguments{
\item{m}{an \code{MAF} object}
}
\description{
Generates an object of class \code{MultiAssayExperiment} from \code{MAF} object
}
\examples{
laml.maf = system.file('extdata', 'tcga_laml.maf.gz', package = 'maftools')
laml.clin = system.file('extdata', 'tcga_laml_annot.tsv', package = 'maftools')
laml = read.maf(maf = laml.maf, clinicalData = laml.clin)
maf2mae(laml)
}
