% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maf_stats.R
\name{maf_stats}
\alias{maf_stats}
\title{Basic information about the input MAF file.}
\usage{
maf_stats(maf_file, removeSilent = TRUE)
}
\arguments{
\item{maf_file}{input file in MAF format.}

\item{removeSilent}{logical. Whether to discard silent mutations ("Silent","Intron","RNA","3'UTR"). Default is TRUE.}
}
\value{
returns a list of data frames with number of mutations per sample, mutations classified according to variant type and variant classification.
}
\description{
Takes MAF file as input and gives intial impression of the data, such as number of mutations per sample, variant types, etc. Also plots barplot and boxplot of variant distributions.
}

