\name{mafTools-package}
\alias{mafTools-package}
\alias{mafTools}
\docType{package}
\title{
Basic statistics about MAF file.

}
\description{
Takes MAF file as input and performs various operations such as classfiy each mutation into Transitions and Transversions, estimates no. mutations per sample, type of mutations, etc. 

}
\details{
\tabular{ll}{
Package: \tab mafTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-29\cr
License: \tab MIT\cr
}

}
\author{
Anand MT

Maintainer: Anand MT <anand_mt@hotmail.com>

}
\references{
Mutation Annotation Format: https://wiki.nci.nih.gov/display/TCGA/Mutation+Annotation+Format+%28MAF%29+Specification
Transitions and Transversions: http://www.mun.ca/biology/scarr/Transitions_vs_Transversions.html
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }

