% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntcounts.R
\name{ntcounts}
\alias{ntcounts}
\title{extract nucleotide counts for targeted variants from the BAM file.}
\usage{
ntcounts(
  bam = NULL,
  loci = NULL,
  mapq = 10,
  sam_flag = 1024,
  op = NULL,
  fa = NULL
)
}
\arguments{
\item{bam}{Input bam file. Required.}

\item{loci}{Loci file. First two columns should contain chromosome and position (1-based)}

\item{mapq}{Map quality. Default 10}

\item{sam_flag}{SAM FLAG to filter reads. Default 1024}

\item{op}{Output file basename. Default parses from BAM file}

\item{fa}{Indexed fasta file. If provided, extracts and adds reference base to the output tsv.}
}
\description{
Given a BAM file and target loci, `ntcounts` fetches redcounts for A, T, G, C, Ins, and Del.
}
