% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoPathway.R
\name{OncogenicPathways}
\alias{OncogenicPathways}
\title{Enrichment of known oncogenic pathways}
\usage{
OncogenicPathways(maf, pathways = NULL, fontSize = 1, panelWidths = c(2, 4, 4))
}
\arguments{
\item{maf}{an \code{\link{MAF}} object generated by \code{\link{read.maf}}}

\item{pathways}{Can be a two column data.frame/tsv-file with pathway-name and genes involved in them. Default `NULL`, uses a predefined list of pathways. See reference for details.}

\item{fontSize}{Default 1}

\item{panelWidths}{Default c(2, 4, 4)}
}
\value{
Prints fraction of altered pathway
}
\description{
Checks for enrichment of known oncogenic pathways
}
\details{
Oncogenic signalling pathways are derived from TCGA cohorts. See reference for details.
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
OncogenicPathways(maf = laml)
}
\references{
Sanchez-Vega F, Mina M, Armenia J, Chatila WK, Luna A, La KC, Dimitriadoy S, Liu DL, Kantheti HS, Saghafinia S et al. 2018. Oncogenic Signaling Pathways in The Cancer Genome Atlas. Cell 173: 321-337 e310
}
\seealso{
\code{\link{PlotOncogenicPathways}}
}
