% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModOncoplot.R
\name{oncoplot.cna}
\alias{oncoplot.cna}
\title{oncoplot.cna}
\usage{
oncoplot.cna(maf, cna, top = 20, genes = NULL, mutsig = NULL,
  mutsigQval = 0.1, drawRowBar = TRUE, drawColBar = TRUE,
  clinicalFeatures = NULL, annotationDat = NULL,
  annotationColor = NULL, genesToIgnore = NULL,
  showTumorSampleBarcodes = FALSE, removeNonMutated = TRUE,
  colors = NULL, sortByMutation = FALSE, sortByAnnotation = FALSE,
  annotationOrder = NULL, keepGeneOrder = FALSE,
  GeneOrderSort = TRUE, sampleOrder = NULL, writeMatrix = FALSE,
  fontSize = 10, SampleNamefontSize = 10, titleFontSize = 15,
  legendFontSize = 12, annotationFontSize = 12,
  annotationTitleFontSize = 12)
}
\description{
modified oncoplot function
}
\examples{

}
