% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{oncoplot}
\alias{oncoplot}
\title{draw an oncoplot}
\usage{
oncoplot(maf, top = 20, genes = NULL, mutsig = NULL,
  mutsigQval = 0.1, drawRowBar = TRUE, drawColBar = TRUE,
  includeColBarCN = TRUE, draw_titv = FALSE, logColBar = FALSE,
  clinicalFeatures = NULL, exprsTbl = NULL, additionalFeature = NULL,
  additionalFeaturePch = 20, additionalFeatureCol = "white",
  additionalFeatureCex = 0.9, annotationDat = NULL,
  annotationColor = NULL, genesToIgnore = NULL,
  showTumorSampleBarcodes = FALSE, barcode_mar = 4, gene_mar = 5,
  removeNonMutated = TRUE, fill = TRUE, cohortSize = NULL,
  colors = NULL, sortByMutation = FALSE, sortByAnnotation = FALSE,
  numericAnnoCol = NULL, groupAnnotationBySize = TRUE,
  annotationOrder = NULL, keepGeneOrder = FALSE,
  GeneOrderSort = TRUE, sampleOrder = NULL, writeMatrix = FALSE,
  sepwd_genes = 0.5, sepwd_samples = 0.25, fontSize = 0.8,
  SampleNamefontSize = 1, showTitle = TRUE, titleFontSize = 1.5,
  legendFontSize = 1.2, annotationFontSize = 1.2, bgCol = "#CCCCCC",
  borderCol = "white", colbar_pathway = FALSE)
}
\arguments{
\item{maf}{an \code{\link{MAF}} object generated by \code{\link{read.maf}}}

\item{top}{how many top genes to be drawn. defaults to 20.}

\item{genes}{Just draw oncoplot for these genes. Default NULL.}

\item{mutsig}{Mutsig resuts if availbale. Usually file named sig_genes.txt If provided plots significant genes and correpsonding Q-values as side row-bar. Default NULL.}

\item{mutsigQval}{Q-value to choose significant genes from mutsig results. Default 0.1}

\item{drawRowBar}{logical plots barplot for each gene. Default \code{TRUE}.}

\item{drawColBar}{logical plots barplot for each sample. Default \code{TRUE}.}

\item{includeColBarCN}{Whether to include CN in column bar plot. Default TRUE}

\item{draw_titv}{logical Includes TiTv plot. \code{FALSE}}

\item{logColBar}{Plot top bar plot on log10 scale. Default \code{FALSE}.}

\item{clinicalFeatures}{columns names from `clinical.data` slot of \code{MAF} to be drawn in the plot. Dafault NULL.}

\item{exprsTbl}{Expression values if availble. Must be a data.frame with two columns containing to gene names and expression values.}

\item{additionalFeature}{a vector of length two indicating column name in the MAF and the factor level to be highlighted.}

\item{additionalFeaturePch}{Default 20}

\item{additionalFeatureCol}{Default "white"}

\item{additionalFeatureCex}{Default 0.9}

\item{annotationDat}{If MAF file was read without clinical data, provide a custom \code{data.frame} with a column \code{Tumor_Sample_Barcode} containing sample names along with rest of columns with annotations.
You can specify which columns to be drawn using `clinicalFeatures` argument.}

\item{annotationColor}{Custom colors to use for `clinicalFeatures`. Must be a named list containing a named vector of colors. Default NULL. See example for more info.}

\item{genesToIgnore}{do not show these genes in Oncoplot. Default NULL.}

\item{showTumorSampleBarcodes}{logical to include sample names.}

\item{barcode_mar}{Default 4}

\item{gene_mar}{Default 5}

\item{removeNonMutated}{Logical. If \code{TRUE} removes samples with no mutations in the oncoplot for better visualization. Default \code{TRUE}.}

\item{fill}{Logical. If \code{TRUE} draws genes and samples as blank grids even when they are not altered.}

\item{cohortSize}{Number of sequenced samples in the cohort. Default all samples from Cohort. You can manually specify the cohort size. Default \code{NULL}}

\item{colors}{named vector of colors for each Variant_Classification.}

\item{sortByMutation}{Force sort matrix according mutations. Helpful in case of MAF was read along with copy number data. Default FALSE.}

\item{sortByAnnotation}{logical sort oncomatrix (samples) by provided `clinicalFeatures`. Sorts based on first `clinicalFeatures`.  Defaults to FALSE. column-sort}

\item{numericAnnoCol}{color palette used for numeric annotations. Default 'YlOrBr' from RColorBrewer}

\item{groupAnnotationBySize}{Further group `sortByAnnotation` orders by their size.  Defaults to TRUE. Largest groups comes first.}

\item{annotationOrder}{Manually specify order for annotations. Works only for first `clinicalFeatures`. Default NULL.}

\item{keepGeneOrder}{logical whether to keep order of given genes. Default FALSE, order according to mutation frequency}

\item{GeneOrderSort}{logical this is applicable when `keepGeneOrder` is TRUE. Default TRUE}

\item{sampleOrder}{Manually speify sample names for oncolplot ordering. Default NULL.}

\item{writeMatrix}{writes character coded matrix used to generate the plot to an output file.}

\item{sepwd_genes}{Default 0.5}

\item{sepwd_samples}{Default 0.25}

\item{fontSize}{font size for gene names. Default 0.8.}

\item{SampleNamefontSize}{font size for sample names. Default 1}

\item{showTitle}{Default TRUE}

\item{titleFontSize}{font size for title. Default 1.5}

\item{legendFontSize}{font size for legend. Default 1.2}

\item{annotationFontSize}{font size for annotations. Default 1.2}

\item{bgCol}{Background grid color for wild-type (not-mutated) samples. Default gray - "#CCCCCC"}

\item{borderCol}{border grid color (not-mutated) samples. Default 'white'.}

\item{colbar_pathway}{Draw top column bar with respect to diplayed pathway. Default FALSE.}
}
\value{
None.
}
\description{
takes output generated by read.maf and draws an oncoplot
}
\details{
Takes maf file as input and plots it as a matrix. Any desired clincal features can be added at the bottom of the oncoplot by providing \code{clinicalFeatures}.
Oncoplot can be sorted either by mutations or by clinicalFeatures using arguments \code{sortByMutation} and \code{sortByAnnotation} respectively.
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml.clin = system.file('extdata', 'tcga_laml_annot.tsv', package = 'maftools')
laml <- read.maf(maf = laml.maf, clinicalData = laml.clin)
#Basic onocplot
oncoplot(maf = laml, top = 3)
#Changing colors for variant classifications (You can use any colors, here in this example we will use a color palette from RColorBrewer)
col = RColorBrewer::brewer.pal(n = 8, name = 'Paired')
names(col) = c('Frame_Shift_Del','Missense_Mutation', 'Nonsense_Mutation', 'Multi_Hit', 'Frame_Shift_Ins',
               'In_Frame_Ins', 'Splice_Site', 'In_Frame_Del')
#Color coding for FAB classification; try getAnnotations(x = laml) to see available annotations.
fabcolors = RColorBrewer::brewer.pal(n = 8,name = 'Spectral')
names(fabcolors) = c("M0", "M1", "M2", "M3", "M4", "M5", "M6", "M7")
fabcolors = list(FAB_classification = fabcolors)
oncoplot(maf = laml, colors = col, clinicalFeatures = 'FAB_classification', sortByAnnotation = TRUE, annotationColor = fabcolors)
}
\seealso{
\code{\link{oncostrip}}
}
