% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint.R
\name{oncoprint}
\alias{oncoprint}
\title{draw an oncoprint similar to the cBioportal output.}
\usage{
oncoprint(maf, genes = NULL, sort = T, legend = T, annotation = NULL,
  removeNonMutated = F, top = 5, showTumorSampleBarcodes = FALSE)
}
\arguments{
\item{maf}{output generated by \code{read.maf}}

\item{genes}{draw oncoprint for these genes. default NULL. Plots top 5 genes.}

\item{sort}{logical sort oncomatrix for enhanced visualization. Defaults to True.}

\item{top}{how many top genes to be drawn. defaults to 20.}

\item{showTumorSampleBarcodes}{logical to include sample names.}
}
\description{
draw an oncoprint similar to the cBioportal output.
}
\details{
takes output generated by \code{read.maf} and draw oncoprint using ggplot2.
}

