% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncostrip.R
\name{oncostrip}
\alias{oncostrip}
\title{draw an oncostrip similar to cBioportal oncoprinter output.}
\usage{
oncostrip(maf, genes = NULL, top = 5, colors = NULL, sort = TRUE,
  annotationCols = NULL, annotationDat = NULL, sortByAnnotation = FALSE,
  removeNonMutated = TRUE, showTumorSampleBarcodes = FALSE,
  annotationColor = NULL)
}
\arguments{
\item{maf}{an \code{\link{MAF}} object generated by \code{read.maf}}

\item{genes}{draw oncoprint for these genes. default NULL. Plots top 5 genes.}

\item{top}{how many top genes to be drawn. defaults to 5.}

\item{colors}{named vector of colors for each Variant_Classification.}

\item{sort}{logical sort oncomatrix for enhanced visualization. Defaults to TRUE.}

\item{annotationCols}{columns names from `annotation` slot of \code{MAF} to be drawn in the plot. Dafault NULL.}

\item{annotationDat}{If MAF file was read without annotations, provide a custom \code{data.frame} with a column containing Tumor_Sample_Barcodes along with rest of columns with annotations.
You can specify which columns to be drawn using `annotationCols` argument.}

\item{sortByAnnotation}{logical sort oncomatrix by provided annotations. Defaults to FALSE. This is mutually exclusive with \code{sort}.}

\item{removeNonMutated}{Logical. If \code{TRUE} removes samples with no mutations in the oncoplot for better visualization. Default TRUE.}

\item{showTumorSampleBarcodes}{logical to include sample names.}

\item{annotationColor}{list of colors to use for annotation. Default NULL.}
}
\value{
None.
}
\description{
draw an oncostrip similar to cBioportal oncoprinter output.
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
dev.new()
oncostrip(maf = laml, genes = c('NPM1', 'RUNX1'))

}
\seealso{
\code{\link{oncoplot}}
}
