% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mostdepth.R
\name{plot_mosdepth_t}
\alias{plot_mosdepth_t}
\title{Plot results from mosdepth output}
\usage{
plot_mosdepth_t(bed = NULL, col = c("#95a5a6", "#7f8c8d"))
}
\arguments{
\item{bed}{mosdepth output}

\item{col}{Colors. Default c("#95a5a6", "#7f8c8d")}
}
\description{
Plot results from mosdepth output
}
