% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotApobecDiff.R
\name{plotApobecDiff}
\alias{plotApobecDiff}
\title{Plot differences between APOBEC enriched and non-APOBEC enriched samples.}
\usage{
plotApobecDiff(tnm, maf)
}
\arguments{
\item{tnm}{output generated by \code{\link{trinucleotideMatrix}}}

\item{maf}{an \code{\link{MAF}} object used to generate the matrix}
}
\value{
list of table containing differenatially altered genes. This can be passed to \code{\link{forestPlot}} to plot results.
}
\description{
Plots differences between APOBEC enriched and non-APOBEC enriched samples
}
\details{
Plots differences between APOBEC enriched and non-APOBEC enriched samples (TCW). Plot includes differences in mutations load, tCw motif distribution and top genes altered.
}
\examples{
\dontrun{
laml.tnm <- trinucleotideMatrix(maf = laml, ref_genome = 'hg19.fa', prefix = 'chr',
add = TRUE, useSyn = TRUE)
plotApobecDiff(laml.tnm)
}
}
\seealso{
\code{\link{trinucleotideMatrix}} \code{\link{plotSignatures}}
}
