% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cbs_segs.R
\name{plotCBSsegments}
\alias{plotCBSsegments}
\title{Plots segmented copy number data.}
\usage{
plotCBSsegments(cbsFile = NULL, maf = NULL, tsb = NULL, chr = NULL,
  savePlot = FALSE, width = 6, height = 3, labelAll = FALSE,
  genes = NULL)
}
\arguments{
\item{cbsFile}{CBS segmented copy number file. Column names should be Sample, Chromosome, Start, End, Num_Probes and Segment_Mean (log2 scale).}

\item{maf}{optional \code{\link{MAF}}}

\item{tsb}{If segmentation file contains many samples (as in gistic input), specify sample name here. Defualt plots all samples. Make sure sample names in
Sample column matches to those in Tumor_Sample_Barcodes of MAF.}

\item{chr}{Just plot this chromosome.}

\item{savePlot}{If true plot is saved as pdf.}

\item{width}{width of plot}

\item{height}{height of plot}

\item{labelAll}{If true and if maf object is specified, maps all mutataions from maf onto segments. Default FALSE, maps only variants on copy number altered regions.}

\item{genes}{highlight only these variants}
}
\value{
ggplot object
}
\description{
Plots segmented copy number data.
}
\details{
this function takes segmented copy number data and plots it. If MAF object is specified, all mutations are highlighted on the plot.
}
\examples{
tcga.ab.009.seg <- system.file("extdata", "TCGA.AB.3009.hg19.seg.txt", package = "maftools")
plotCBSsegments(cbsFile = tcga.ab.009.seg)

}

