% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCophenetic.R
\name{plotCophenetic}
\alias{plotCophenetic}
\title{Draw an elbow plot of cophenetic correlation metric.}
\usage{
plotCophenetic(res = NULL, bestFit = NULL)
}
\arguments{
\item{res}{output from \code{\link{estimateSignatures}}}

\item{bestFit}{rank to highlight. Default NULL}
}
\description{
Draw an elbow plot of cophenetic correlation metric.
}
\details{
This function draws an elbow plot of cophenetic correlation metric.
}
\seealso{
\code{\link{estimateSignatures}} \code{\link{plotCophenetic}}
}
