% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCopheneticMetric.R
\name{plotCopheneticMetric}
\alias{plotCopheneticMetric}
\title{Plot elbow plot for Cophenetic Metric against range of factorization ranks}
\usage{
plotCopheneticMetric(nmfRes, bestFit = "auto")
}
\arguments{
\item{nmfRes}{Signature results from \code{\link{extractSignatures}}}

\item{bestFit}{Choice of best fit. Can be "auto", or an integer. Default "auto"}
}
\description{
Plot elbow plot for Cophenetic Metric against range of factorization ranks
}
\details{
Use this function to plot Cophenetic correlation metric against a range of factorization ranks tried. This plot helps to choose optimal factorization rank.
Ideally optimal rank is the one at which Cophenetic metric significantly drops, and after which there is little change in the values (elbow).
}
