% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClinicalEnrichment.R
\name{plotEnrichmentResults}
\alias{plotEnrichmentResults}
\title{Plots results from clinicalEnrichment analysis}
\usage{
plotEnrichmentResults(
  enrich_res,
  pVal = 0.05,
  ORthr = 1,
  featureLvls = NULL,
  cols = NULL,
  annoFontSize = 0.8,
  geneFontSize = 0.8,
  legendFontSize = 0.8,
  showTitle = TRUE
)
}
\arguments{
\item{enrich_res}{results from \code{\link{clinicalEnrichment}} or \code{\link{signatureEnrichment}}}

\item{pVal}{Default 0.05}

\item{ORthr}{Default 1. Odds ratio threshold. >1 indicates positive enrichment in the group of interest.}

\item{featureLvls}{Plot results from the selected levels. Default NULL, plots all.}

\item{cols}{named vector of colors for factor in a clinical feature. Default NULL}

\item{annoFontSize}{cex for annotation font size. Default 0.8}

\item{geneFontSize}{cex for gene font size. Default 0.8}

\item{legendFontSize}{cex for legend font size. Default 0.8}

\item{showTitle}{Default TRUE}
}
\value{
returns nothing.
}
\description{
Plots results from clinicalEnrichment analysis
}
\seealso{
\code{\link{clinicalEnrichment}} \code{\link{signatureEnrichment}}
}
