% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMafSummary.R
\name{plotmafSummary}
\alias{plotmafSummary}
\title{Plots maf summary.}
\usage{
plotmafSummary(maf, file = NULL, width = 6, height = 5)
}
\arguments{
\item{file}{If given pdf file will be generated.}

\item{width}{plot parameter for output file.}

\item{height}{plot parameter for output file.}

\item{an}{MAF object generated by \code{read.maf}}
}
\description{
Plots maf summary.
}

