% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOncodriveClust.R
\name{plotOncodrive}
\alias{plotOncodrive}
\title{Plots results from \code{oncodrive}}
\usage{
plotOncodrive(res = NULL, fdrCutOff = 0.05, useFraction = F)
}
\arguments{
\item{res}{results from \code{oncodrive}}

\item{fdrCutOff}{fdr cutoff to call a gene as a driver.}

\item{useFraction}{if TRUE uses a fraction of total variants as X-axis scale instead of absolute counts.}
}
\value{
a ggplot object which can be further modified.
}
\description{
Takes results from \code{oncodrive} and plots them as a scatter plot. Size of the gene shows number of clusters (hotspots), x-axis can either be an absolute number of variants
accumulated in these clusters or a fraction of total variants found in these clusters. y-axis is fdr values transformed into -log10 for better representation. Lables indicate Gene name with number clusters
observed.
}

