% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPathways.R
\name{PlotOncogenicPathways}
\alias{PlotOncogenicPathways}
\title{Plot oncogenic pathways}
\usage{
PlotOncogenicPathways(maf, pathways = NULL, fullPathway = FALSE,
  removeNonMutated = TRUE, tsgCol = "red", ogCol = "royalblue",
  fontSize = 0.6)
}
\arguments{
\item{maf}{an \code{\link{MAF}} object generated by \code{\link{read.maf}}}

\item{pathways}{Name of pathways to be drawn}

\item{fullPathway}{Include all genes from the pathway. Defaulr FALS only plots mutated genes}

\item{removeNonMutated}{Default TRUE}

\item{tsgCol}{Color for tumro suppressor genes. Default red}

\item{ogCol}{Color for onco genes. Default royalblue}

\item{fontSize}{Default 0.6}
}
\description{
Plot oncogenic pathways
}
\details{
Draws oncoplot of oncogenic pathway.
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
PlotOncogenicPathways(maf = laml, pathways = "RTK-RAS")
}
\references{
Sanchez-Vega F, Mina M, Armenia J, Chatila WK, Luna A, La KC, Dimitriadoy S, Liu DL, Kantheti HS, Saghafinia S et al. 2018. Oncogenic Signaling Pathways in The Cancer Genome Atlas. Cell 173: 321-337 e310
}
\seealso{
\code{\link{OncogenicPathways}}
}
