% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vaf.R
\name{plotVaf}
\alias{plotVaf}
\title{Plots vaf distribution of genes}
\usage{
plotVaf(maf, vafCol = NULL, genes = NULL, density = F, violin = F,
  top = 5)
}
\arguments{
\item{maf}{output generated by \code{read.maf}}

\item{vafCol}{manually specify column name for vafs. Default looks for column 't_vaf'}

\item{genes}{specify genes for which plots has to be generated}

\item{density}{logical whether to plot density plot of vaf}

\item{violin}{plots violin plot}

\item{top}{if \code{genes} is NULL plots top n number of genes. Defaults to 5.}
}
\value{
ggplot object which can be further modified.
}
\description{
takes output generated by read.maf and plots vaf distribution.
}

