% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepAscat.R
\name{prep_ascat}
\alias{prep_ascat}
\title{Prepare input files for ASCAT}
\usage{
prep_ascat(
  t_counts = NULL,
  n_counts = NULL,
  sample_name = NA,
  min_depth = 30,
  normalize = FALSE
)
}
\arguments{
\item{t_counts}{read counts from tumor generated by `get_counts`}

\item{n_counts}{read counts from normal generated by `get_counts`}

\item{sample_name}{Sample name. Used as a basename for output files. Default NA, parses from t_counts file.}

\item{min_depth}{Min read depth required to consider a marker. Default 30}

\item{normalize}{If TRUE, normalizes for library size}
}
\value{
An \code{\link{ascat.loadData}} object; ascat data structure
}
\description{
Prepare input files for ASCAT
}
