% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepAscat_t.R
\name{prep_ascat_t}
\alias{prep_ascat_t}
\title{Prepare input files for ASCAT tumor only samples}
\usage{
prep_ascat_t(
  t_counts = NULL,
  sample_name = NA,
  min_depth = 30,
  PLATFORM = "AffySNP6"
)
}
\arguments{
\item{t_counts}{read counts from tumor generated by `get_counts`}

\item{sample_name}{Sample name. Used as a basename for output files. Default NA, parses from t_counts file.}

\item{min_depth}{Min read depth required to consider a marker. Default 30}

\item{PLATFORM}{Default AffySNP6. Only change if you have used custom loci. See here for available options https://www.crick.ac.uk/research/labs/peter-van-loo/software}
}
\value{
An \code{\link{ascat.loadData}} object; ascat data structure
}
\description{
Prepare input files for ASCAT tumor only samples
}
