% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinMap.R
\name{proteinMap}
\alias{proteinMap}
\title{Draws lollipop plot of amino acid changes on to Protein structure.}
\usage{
proteinMap(maf, gene = NULL, label = F)
}
\arguments{
\item{maf}{maf maf output generated by \code{read.maf}}

\item{gene}{HGNC symbol for which protein structure to be drawn.}

\item{label}{if true lables dots with amino acid conversion using ggrepel package.}
}
\value{
ggplot object of the plot, which can be futher modified.
}
\description{
Draws lollipop plot of amino acid changes.
}
\details{
This function looks for field "AAChange" in maf file. Since there is no hard rule for filed name for amino acid changes, it is advised to manually rename column name to "AAChange" in MAF file if its different.
Protein Domain information is obtained from UniProt.
}

