% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfallPlot.R
\name{rainfallPlot}
\alias{rainfallPlot}
\title{Rainfall plot to display kataegis or hyper mutated genomic regions.}
\usage{
rainfallPlot(maf, tsb = NULL, color = NULL, savePlot = FALSE, width = 6,
  height = 3)
}
\arguments{
\item{maf}{an \code{\link{MAF}} object generated by \code{\link{read.maf}}. Required.}

\item{tsb}{specify sample names (Tumor_Sample_Barcodes) for which plotting has to be done. If NULL, draws plot for most mutated sample.}

\item{color}{named vector of colors for each coversion class.}

\item{savePlot}{If TRUE plot is saved to output pdf. Default FALSE.}

\item{width}{width of plot to be saved.}

\item{height}{height of plot to be saved.}
}
\value{
returns ggplot object of the plot which can be further modified.
}
\description{
Plots inter variant distance as a function of genomic locus.
}

