% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_logR.R
\name{segment_logR}
\alias{segment_logR}
\title{Segment log ratio values with DNACopy}
\usage{
segment_logR(tumor_logR = NULL, sample_name = "tumor")
}
\arguments{
\item{tumor_logR}{sample.tumour.logR.txt file generated by \code{\link{prep_ascat}}}

\item{sample_name}{Default "tumor"}
}
\value{
Invisibly returns \code{\link{DNAcopy}} object
}
\description{
The function takes tumor logR file generated by \code{\link{prep_ascat}} and performs segmentation with \code{\link{DNAcopy}}
}
