% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_logR.R
\name{segmentLogR}
\alias{segmentLogR}
\title{Segment and plot log ratio values with DNACopy}
\usage{
segmentLogR(tumor_logR = NULL, sample_name = NULL, build = "hg19", ...)
}
\arguments{
\item{tumor_logR}{logR.txt file generated by \code{\link{prep_ascat}} or \code{\link{prep_ascat_t}}}

\item{sample_name}{Default NULL. Parses from `tumor_logR` file}

\item{build}{Reference genome. Default hg19. Can be hg18, hg19, or hg38}
}
\value{
Invisibly returns \code{\link{DNAcopy}} object
}
\description{
The function takes logR file generated by \code{\link{prep_ascat}} or \code{\link{prep_ascat_t}} and performs segmentation with \code{\link{DNAcopy}}
}
\seealso{
\code{\link{gtSNPs}} \code{\link{prep_ascat}}
}
