% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMaf.R
\name{setMaf}
\alias{setMaf}
\alias{setdiffMAF}
\alias{unionMAF}
\alias{intersectMAF}
\title{Set Operations for MAF objects}
\usage{
setdiffMAF(x, y, mafObj = TRUE, ...)

unionMAF(x, y, mafObj = TRUE, ...)

intersectMAF(x, y, mafObj = TRUE, ...)
}
\arguments{
\item{x}{the first `MAF` object.}

\item{y}{the second `MAF` object.}

\item{...}{other parameters passing to `subsetMaf` for subsetting operations.}
}
\description{
Set Operations for MAF objects
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
x <- subsetMaf(maf = laml, tsb = c('TCGA-AB-3009'))
y <- subsetMaf(maf = laml, tsb = c('TCGA-AB-2933'))
setdiffMAF(x, y)
intersectMAF(x, y)
# This is similar to merge_mafs()
unionMAF(x, y)
}
