% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaLoad.R
\name{tcgaAvailable}
\alias{tcgaAvailable}
\title{Prints available TCGA datasets}
\usage{
tcgaAvailable(repo = c("github", "gitee"))
}
\arguments{
\item{repo}{one of "github" (default) and "gitee" (for users in China).}
}
\description{
Prints available TCGA cohorts
}
\examples{
tcgaAvailable()
}
\seealso{
\code{\link{tcgaLoad}}
}
