% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TiTv_stats.R
\name{TiTv_stats}
\alias{TiTv_stats}
\title{Classify SNVs into Transitions and Tranversions.}
\usage{
TiTv_stats(maf_file, is_maf = TRUE, plot = TRUE, header = TRUE)
}
\arguments{
\item{maf_file}{input file either in MAF format or six column format.}

\item{is_maf}{logical. Whether input is in MAF or not. Default is TRUE.}

\item{plot}{logical. Plots distribution of various conversion events as a boxplot.}

\item{header}{logical. Whether input file has a header line. Default is TRUE.}
}
\value{
returns a list of dataframes containing raw counts classified into six different classes and fraction of conversion for each sample.
}
\description{
Takes MAF file or simple six column file with chr, start, end, ref_allele, alt_allele and sample id as input
 and classifies them into six classes of converions.
}

