% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{tmb}
\alias{tmb}
\title{Estimate Tumor Mutation Burden}
\usage{
tmb(maf, captureSize = 50, logScale = TRUE)
}
\arguments{
\item{maf}{maf \code{\link{MAF}} object}

\item{captureSize}{capture size for input MAF in MBs. Default 50MB.}

\item{logScale}{Default TRUE. For plotting purpose only.}
}
\value{
data.table with TMB for every sample
}
\description{
Estimates Tumor Mutation Burden in terms of per megabases
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
tmb(maf = laml)
}
