% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrinucleotideMatrix.R
\name{trinucleotideMatrix}
\alias{trinucleotideMatrix}
\title{Extract single 5' and 3' bases flanking the mutated site.}
\usage{
trinucleotideMatrix(maf, ref_genome, prefix = NULL, add = TRUE,
  ignoreChr = NULL)
}
\arguments{
\item{maf}{an MAF object generated by \code{\link{read.maf}}}

\item{ref_genome}{faidx indexed refrence fasta file.}

\item{prefix}{Prefix to add or remove from contig names in MAF file.}

\item{add}{If prefix is used, default is to add prefix to contig names in MAF file. If false prefix will be removed from contig names.}

\item{ignoreChr}{Chromsomes to remove from analysis. e.g. chrM}
}
\value{
A matrix of dimension nx96, where n is the number of samples in the MAF.
}
\description{
Extract single 5' and 3' bases flanking the mutated site.
}
\details{
Extracts immediate 5' and 3' bases flanking the mutated site and classifies them into 96 substitution classes.
This function loads reference genome into memeory. Typical human geneome occupies a peak memory of ~3 gb while extracting bases.
}

