% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr.R
\name{vcr}
\alias{vcr}
\title{Samll internal function to make complex events. Ignore this.}
\usage{
vcr(x, gis = FALSE)
}
\arguments{
\item{x}{charcter to split}

\item{gis}{Is input from gistic. Logical.}
}
\description{
Samll internal function to make complex events. Ignore this.
}

