% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrteMafSummary.R
\name{write.mafSummary}
\alias{write.mafSummary}
\title{Writes maf summaries to an output tab-delimited text file.}
\usage{
write.mafSummary(maf, basename = NULL)
}
\arguments{
\item{maf}{an MAF object generated by \code{read.maf}}

\item{basename}{basename for out file to be written.}
}
\description{
Writes maf summaries to an output tab-delimited text file.
}

