% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coGisticChromPlotV.R
\name{yload_gistic}
\alias{yload_gistic}
\title{Read the gscore and broad data from a GISTIC2.0 output folder}
\usage{
yload_gistic(gistic_res_dir)
}
\arguments{
\item{gistic_res_dir}{the path to the GISTIC2.0 output folder}
}
\value{
LoadedGisticObj object which contains a GISTIC object and a broad data table
}
\description{
Read the gscore and broad data from a GISTIC2.0 output folder
}
\examples{
gistic_res_folder = system.file("extdata",package = "maftools")
laml.gistic = yload_gistic(gistic_res_folder)

}
