% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coOncoplot.R
\name{coOncoplot}
\alias{coOncoplot}
\title{Draw two oncoplots side by side for cohort comparision.}
\usage{
coOncoplot(m1, m2, genes = NULL, colors = NULL, removeNonMutated = TRUE,
  m1Name = NULL, m2Name = NULL)
}
\arguments{
\item{m1}{first \code{\link{MAF}} object}

\item{m2}{second \code{\link{MAF}} object}

\item{genes}{draw these genes. Default plots top 5 mutated genes from two cohorts.}

\item{colors}{named vector of colors for each Variant_Classification.}

\item{removeNonMutated}{Logical. If \code{TRUE} removes samples with no mutations in the oncoplot for better visualization. Default \code{TRUE}.}

\item{m1Name}{optional name for first cohort}

\item{m2Name}{optional name for second cohort}
}
\value{
Returns nothing. Just draws plot.
}
\description{
Draw two oncoplots side by side for cohort comparision.
}
\details{
Draws two oncoplots side by side to display difference between two cohorts.
}
\examples{
#' ##Primary and Relapse APL
primary.apl <- system.file("extdata", "APL_primary.maf.gz", package = "maftools")
relapse.apl <- system.file("extdata", "APL_relapse.maf.gz", package = "maftools")
##Read mafs
primary.apl <- read.maf(maf = primary.apl)
relapse.apl <- read.maf(maf = relapse.apl)
##Plot
coOncoplot(m1 = primary.apl, m2 = relapse.apl, m1Name = 'Primary APL', m2Name = 'Relapse APL')
dev.off()
}

