% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSignatures.R
\name{estimateSignatures}
\alias{estimateSignatures}
\title{Estimate number of signatures based on cophenetic correlation metric}
\usage{
estimateSignatures(mat, nTry = 6, parallel = 4, pConstant = NULL,
  plotBestFitRes = FALSE)
}
\arguments{
\item{mat}{Input matrix of diemnsion nx96 generated by \code{\link{trinucleotideMatrix}}}

\item{nTry}{Maximum number of signatures to try. Default 6.}

\item{parallel}{Default 4. Number of cores to use. See note on \code{\link{nmf}} for MAC users.}

\item{pConstant}{A small positive value to add to the matrix. Use it ONLY if the functions throws an \code{non-conformable arrays} error}

\item{plotBestFitRes}{plots consensus heatmap for range of values tried. Default FALSE}
}
\value{
a list with \code{NMF.rank} object and summary stats.
}
\description{
Estimate number of signatures based on cophenetic correlation metric
}
\details{
This function decomposes a non-negative matrix into n signatures.
Extracted signatures are compared against 30 experimentally validated signatures by calculating cosine similarity. See http://cancer.sanger.ac.uk/cosmic/signatures for details.
}
\examples{
\dontrun{
laml.tnm <- trinucleotideMatrix(maf = laml, ref_genome = 'hg19.fa', prefix = 'chr',
add = TRUE, useSyn = TRUE)
laml.sign <- estimateSignatures(mat = laml.tnm, plotBestFitRes = FALSE)
}
}
\seealso{
\code{\link{plotCophenetic}} \code{\link{extractSignatures}} \code{\link{trinucleotideMatrix}}
}
