% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBases.R
\name{ExtractSignatures}
\alias{ExtractSignatures}
\title{Extract mutational signatures from trinucletide context.}
\usage{
ExtractSignatures(maf, ref_genome, prefix = NULL, add = TRUE,
  ignoreChr = NULL, n = 3)
}
\arguments{
\item{maf}{output generated by \code{read.maf}}

\item{ref_genome}{faidx indexed refrence fasta file.}

\item{prefix}{Prefix to add or remove from contig names in MAF file.}

\item{add}{If prefix is used, default is to add prefix to contig names in MAF file. If false prefix will be removed from contig names.}

\item{ignoreChr}{Chromsomes to remove from analysis. e.g. chrM}

\item{n}{decompose matrix into n signatures. Default 3 signatures.}
}
\value{
returns a list with signatures and 96 substitution matrix.
}
\description{
Extracts adjacent bases to mutation site and classifies them into 96 substitution classes based on trinucleotide context followed by NMF decomposition to extract \code{n} signatures.
}
\details{
This function loads reference genome into memeory. Typical human geneome occupies a peak memory of ~3 gb while extracting bases. NMF decomposition uses \code{nmf} function from NMF package.
Extracted signatures are compared against 21 experimentally validated signatures by calculating correlation coefficient. See http://www.nature.com/nature/journal/v500/n7463/fig_tab/nature12477_F2.html for details.
}

