% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBases.R
\name{extractSignatures}
\alias{extractSignatures}
\title{Extract mutational signatures from trinucletide context.}
\usage{
extractSignatures(mat, n = NULL)
}
\arguments{
\item{mat}{Input matrix of diemnsion nx96 generated by \code{trinucleotideMatrix}}

\item{n}{decompose matrix into n signatures. Default NULL. Tries to predict best one by running NMF on a range of values and chooses the one with highest cophenetic correlation coefficient.}
}
\value{
returns a list with decomposed signatures and correlatoon table against validated signatures.
}
\description{
Decompose a matrix of mutations classified into 96 substitution classes into \code{n} signatures.
}
\details{
This function decomposes a non-negative matrix into
Extracted signatures are compared against 21 experimentally validated signatures by calculating correlation coefficient. See http://www.nature.com/nature/journal/v500/n7463/fig_tab/nature12477_F2.html for details.
}

