% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlot.R
\name{forestPlot}
\alias{forestPlot}
\title{Draw forest plot for differences betweeen cohorts.}
\usage{
forestPlot(mafCompareRes, pVal = 0.05, show = NULL, color = NULL,
  file = NULL, width = 5, height = 6)
}
\arguments{
\item{pVal}{p-value threshold. Default 0.05.}

\item{show}{can be either \code{stat} or \code{pval}}

\item{color}{vector of colors for plot. Default NULL.}

\item{file}{basename for output file. Plot will sabe to output pdf.}

\item{width}{width of plot to be generated}

\item{height}{height of plot to be generated}

\item{mafComareRes}{results from \code{\link{mafCompare}}}
}
\value{
ggplot object of the plot.
}
\description{
Draw forest plot for differences betweeen cohorts.
}
\details{
Plots results from mafCompare as a forest plot with x-axis log10 converted odds ratio and differentially mutated genes on y-axis.
}
\seealso{
\code{\link{mafCompare}}
}

