% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesToBarcodes.R
\name{genesToBarcodes}
\alias{genesToBarcodes}
\title{Extracts Tumor Sample Barcodes where the given genes are mutated.}
\usage{
genesToBarcodes(maf, genes = NULL, justNames = F)
}
\arguments{
\item{maf}{an MAF object generated by \code{read.maf}}

\item{genes}{Hogo_Symbol for which sample names to be extracted.}

\item{justNames}{if TRUE, just returns samples names instead of summarized tables.}
}
\value{
list of \code{data.table}s with samples in which given genes are mutated.
}
\description{
Extracts Tumor Sample Barcodes where the given genes are mutated.
}

