% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_methods.R
\docType{methods}
\name{getGeneSummary}
\alias{getGeneSummary}
\alias{getGeneSummary,MAF-method}
\alias{getGeneSummary}
\alias{getGeneSummary,GISTIC-method}
\alias{getGeneSummary}
\title{extract gene summary from MAF or GISTIC object}
\usage{
getGeneSummary(x)

\S4method{getGeneSummary}{MAF}(x)

\S4method{getGeneSummary}{GISTIC}(x)
}
\arguments{
\item{x}{An object of class MAF or GISTIC}
}
\value{
gene summary table
}
\description{
extract gene summary from MAF or GISTIC object
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf, removeSilent = TRUE, useAll = FALSE)
getGeneSummary(laml)
}
