% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinMap.R
\name{lollipopPlot}
\alias{lollipopPlot}
\title{Draws lollipop plot of amino acid changes on to Protein structure.}
\usage{
lollipopPlot(maf, gene = NULL, label = F, AACol = NULL)
}
\arguments{
\item{maf}{maf maf output generated by \code{read.maf}}

\item{gene}{HGNC symbol for which protein structure to be drawn.}

\item{label}{if true lables dots with amino acid conversion using ggrepel package.}

\item{AACol}{manually specify column name for amino acid changes. Default looks for column 'AAChange'}
}
\value{
ggplot object of the plot, which can be futher modified.
}
\description{
Draws lollipop plot of amino acid changes.
}
\details{
This function by default looks for field "AAChange" in maf file. One can also manually specify field name containing amino acid changes. Protein Domains are retrieved
from Uniprot database.
}

