% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinMap.R
\name{lollipopPlot}
\alias{lollipopPlot}
\title{Draws lollipop plot of amino acid changes on to Protein structure.}
\usage{
lollipopPlot(maf, gene = NULL, refSeqID = NULL, proteinID = NULL,
  labelPos = NULL, AACol = NULL, repel = FALSE)
}
\arguments{
\item{maf}{an MAF object generated by \code{\link{read.maf}}}

\item{gene}{HGNC symbol for which protein structure to be drawn.}

\item{refSeqID}{RefSeq transcript identifier for \code{gene} if known.}

\item{proteinID}{RefSeq protein identifier for \code{gene} if known.}

\item{labelPos}{Amino acid positions to label. If 'all', labels all variants.}

\item{AACol}{manually specify column name for amino acid changes. Default looks for field 'AAChange'}

\item{repel}{If points are too close to each other, use this option to repel them. Default FALSE.}
}
\value{
ggplot object of the plot, which can be futher modified.
}
\description{
Draws lollipop plot of amino acid changes.
}
\details{
This function by default looks for field "AAChange" in maf file. One can also manually specify field name containing amino acid changes.
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf, removeSilent = TRUE, useAll = FALSE)
lollipopPlot(maf = laml, gene = 'KIT', AACol = 'Protein_Change')

}

