% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipopPlot.R
\name{lollipopPlot}
\alias{lollipopPlot}
\title{Draws lollipop plot of amino acid changes on to Protein structure.}
\usage{
lollipopPlot(maf, gene = NULL, refSeqID = NULL, proteinID = NULL,
  labelPos = NULL, AACol = NULL, repel = FALSE, colors = NULL,
  legendTxtSize = 10)
}
\arguments{
\item{maf}{an \code{\link{MAF}} object generated by \code{\link{read.maf}}}

\item{gene}{HGNC symbol for which protein structure to be drawn.}

\item{refSeqID}{RefSeq transcript identifier for \code{gene} if known.}

\item{proteinID}{RefSeq protein identifier for \code{gene} if known.}

\item{labelPos}{Amino acid positions to label. If 'all', labels all variants.}

\item{AACol}{manually specify column name for amino acid changes. Default looks for field 'AAChange'. Changes can be of any format i.e, can be a numeric value or HGVSp annotations (e.g; p.P459L, p.L2195Pfs*30 or p.Leu2195ProfsTer30)}

\item{repel}{If points are too close to each other, use this option to repel them. Default FALSE. Warning: naive method, might make plot ugly in case of too many variants!}

\item{colors}{named vector of colors for each Variant_Classification. Default NULL.}

\item{legendTxtSize}{Text size for legend. Default 10}
}
\value{
ggplot object of the plot, which can be futher modified.
}
\description{
Draws lollipop plot of amino acid changes.
}
\details{
This function by default looks for field "AAChange" in maf file. One can also manually specify field name containing amino acid changes.
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf, removeSilent = TRUE, useAll = FALSE)
lollipopPlot(maf = laml, gene = 'KIT', AACol = 'Protein_Change')

}

