% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafObjects.R
\docType{class}
\name{MAF-class}
\alias{MAF-class}
\alias{MAF}
\title{Class MAF}
\description{
S4 class for storing summarized MAF.
}
\section{Slots}{

\describe{
\item{\code{data}}{data.table of original MAF file.}

\item{\code{variants.per.sample}}{table containing variants per sample}

\item{\code{variant.type.summary}}{table containing variant types per sample}

\item{\code{variant.classification.summary}}{table containing variant classification per sample}

\item{\code{gene.summary}}{table containing variant classification per gene}

\item{\code{oncoMatrix}}{character matrix of dimension n*m where n is number of genes and m is number of variants}

\item{\code{numericMatrix}}{numeric matrix of dimension n*m where n is number of genes and m is number of variants}

\item{\code{summary}}{table with basic MAF summary stats}

\item{\code{classCode}}{mapping between numeric values in numericMatrix and Variant Classification}

\item{\code{maf.silent}}{subset of main MAF containing only silent variants}
}}

\seealso{
\code{\link{getGeneSummary}} \code{\link{getSampleSummary}} \code{\link{getFields}}
}
