% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafCompare.R
\name{mafCompare}
\alias{mafCompare}
\title{compare two cohorts (MAF).}
\usage{
mafCompare(m1, m2, m1Name = NULL, m2Name = NULL, minMut = 5)
}
\arguments{
\item{m1}{first \code{\link{MAF}} object}

\item{m2}{second \code{\link{MAF}} object}

\item{m1Name}{optional name for first cohort}

\item{m2Name}{optional name for second cohort}

\item{minMut}{Consider only genes with minimum this number of samples mutated in atleast one of the cohort for analysis. Helful to ignore single mutated genes. Default 5.}
}
\value{
result list
}
\description{
compare two cohorts (MAF).
}
\details{
Performs fisher test on 2x2 contigency table generated from two cohorts to find differentially mutated genes.
}
\seealso{
\code{\link{forestPlot}}
}

