% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMafs.R
\name{merge_mafs}
\alias{merge_mafs}
\title{Merge multiple mafs into single MAF}
\usage{
merge_mafs(mafs, verbose = TRUE, ...)
}
\arguments{
\item{mafs}{a list of \code{\link{MAF}} objects or data.frames or paths to MAF files.}

\item{verbose}{Default TRUE}

\item{...}{additional arguments passed \code{\link{read.maf}}}
}
\value{
\code{\link{MAF}} object
}
\description{
Merges multiple maf files/objects/data.frames into a single MAF.
}
