% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMafs.R
\name{merge_mafs}
\alias{merge_mafs}
\title{Merge multiple maf files into a single MAF}
\usage{
merge_mafs(mafs, MAFobj = FALSE, ...)
}
\arguments{
\item{mafs}{a vector of maf files.}

\item{MAFobj}{If TRUE, returns result as an \code{\link{MAF}} object. Default FALSE}

\item{...}{additional arguments passed \code{\link{read.maf}}. Only applicable if MAFobj = TRUE.}
}
\value{
data.table of merged MAFs or \code{\link{MAF}} object
}
\description{
Merges multiple maf files into a single MAF by matching column names.
}
