% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oncoPlot.R
\name{oncoPlot}
\alias{oncoPlot}
\title{Plots maf file as a matrix.}
\usage{
oncoPlot(maf_file, removeSilent = TRUE, oncoPlot = TRUE,
  writeMatrix = FALSE)
}
\arguments{
\item{maf_file}{input file in MAF format.}

\item{removeSilent}{logical. Whether to discard silent mutations ("Silent","Intron","RNA","3'UTR") before forming matrix. Default is TRUE.}

\item{oncoPlot}{logical. Whether to plot mutation matrix (Upto top 20 genes). Default is TRUE.}

\item{writeMatrix}{Writes the transformed maf file as a tab delimited text file, to be opened by spreadsheet applications.}
}
\value{
Plots and writes formed matrix to a file.
}
\description{
Takes maf file as input and plots it as a matrix similar to the output generated by oncoprint.
This function heavily depeds on oncoplot script by jokergoo for plotting (https://github.com/jokergoo/ComplexHeatmap/blob/master/vignettes/oncoprint.R)
while taking care of format conversion.
}

