% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{oncoplot}
\alias{oncoplot}
\title{draw an oncoplot}
\usage{
oncoplot(maf, writeMatrix = FALSE, top = 20, bg = "#CCCCCC",
  drawRowBar = T, drawColBar = T, showTumorSampleBarcodes = FALSE,
  annotation = NULL, annotationColor = NULL, genesToIgnore = NULL,
  removeNonMutated = F)
}
\arguments{
\item{maf}{an MAF object generated by \code{\link{read.maf}}}

\item{writeMatrix}{writes oncomatrix to output file.}

\item{top}{how many top genes to be drawn. defaults to 20.}

\item{drawRowBar}{logical plots barplot for each gene.}

\item{drawColBar}{logical plots barplot for each sample.}

\item{showTumorSampleBarcodes}{logical to include sample names.}

\item{annotation}{data.frame with first column containing Tumor_Sample_Barcodes and rest of columns with annotations.}

\item{genesToIgnore}{do not show these genes in Oncoplot. Default NULL.}

\item{removeNonMutated}{Logical. If \code{TRUE} removes samples with no mutations in the oncoplot for better visualization. Default \code{FALSE}.}
}
\description{
takes output generated by read.maf and draws an oncoplot (aka waterfall plot).
}
\details{
Takes maf file as input and plots it as a matrix. Any desired annotations can be added at the bottom of the oncoplot by providing \code{annotation}
}

