% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{oncoplot}
\alias{oncoplot}
\title{draw an oncoplot}
\usage{
oncoplot(maf, top = 20, genes = NULL, mutsig = NULL, mutsigQval = 0.1,
  drawRowBar = TRUE, drawColBar = TRUE, annotationCols = NULL,
  annotationDat = NULL, annotationColor = NULL, genesToIgnore = NULL,
  showTumorSampleBarcodes = FALSE, removeNonMutated = TRUE, colors = NULL,
  fontSize = 10, sortByMutation = FALSE, sortByAnnotation = FALSE,
  writeMatrix = FALSE)
}
\arguments{
\item{maf}{an \code{\link{MAF}} object generated by \code{\link{read.maf}}}

\item{top}{how many top genes to be drawn. defaults to 20.}

\item{genes}{Just draw oncoplot for these genes. Default NULL.}

\item{mutsig}{Mutsig resuts if availbale. Usually file named sig_genes.txt If provided plots significant genes and correpsonding Q-values as side row-bar. Default NULL.}

\item{mutsigQval}{Q-value to choose significant genes from mutsig results. Default 0.1}

\item{drawRowBar}{logical plots barplot for each gene. Default \code{TRUE}.}

\item{drawColBar}{logical plots barplot for each sample. Default \code{TRUE}.}

\item{annotationCols}{columns names from `annotation` slot of \code{MAF} to be drawn in the plot. Dafault NULL.}

\item{annotationDat}{If MAF file was read without annotations, provide a custom \code{data.frame} with a column containing Tumor_Sample_Barcodes along with rest of columns with annotations.
You can specify which columns to be drawn using `annotationCols` argument.}

\item{annotationColor}{list of colors to use for annotation. Default NULL.}

\item{genesToIgnore}{do not show these genes in Oncoplot. Default NULL.}

\item{showTumorSampleBarcodes}{logical to include sample names.}

\item{removeNonMutated}{Logical. If \code{TRUE} removes samples with no mutations in the oncoplot for better visualization. Default \code{TRUE}.}

\item{colors}{named vector of colors for each Variant_Classification.}

\item{fontSize}{font size for gene names. Default 10.}

\item{sortByMutation}{Helpful in case of MAF was read along with copy number data. Default FALSE.}

\item{sortByAnnotation}{logical sort oncomatrix by provided annotations. Defaults to FALSE. This is mutually exclusive with sortByMutation.}

\item{writeMatrix}{writes character coded matrix used to generate the plot to an output file. This can be used as an input
for ComplexHeatmap \link[ComplexHeatmap]{oncoPrint} function if you wish to customize the plot.}
}
\value{
None.
}
\description{
takes output generated by read.maf and draws an oncoplot (aka waterfall plot).
}
\details{
Takes maf file as input and plots it as a matrix. Any desired annotations can be added at the bottom of the oncoplot by providing \code{annotation}.
Oncoplot can be sorted either by mutations or annotations using arguments \code{sortByMutation} and \code{sortByAnnotation} respectively.

Thanks to Ryan Morin for \code{sortByAnnotation} code.
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
oncoplot(maf = laml, top = 3)
}
\seealso{
\code{\link{oncostrip}}
}
