% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncostrip.R
\name{oncostrip}
\alias{oncostrip}
\title{draw an oncostrip similar to cBioportal oncoprinter output.}
\usage{
oncostrip(maf = NULL, ...)
}
\arguments{
\item{maf}{an \code{\link{MAF}} object generated by \code{read.maf}}

\item{...}{arguments passed \code{\link{oncoplot}}}
}
\value{
None.
}
\description{
draw an oncostrip similar to cBioportal oncoprinter output.
}
\details{
This is just a wrapper around  \code{\link{oncoplot}} with drawRowBar and drawColBar set to FALSE
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
dev.new()
oncostrip(maf = laml, genes = c('NPM1', 'RUNX1'))

}
\seealso{
\code{\link{oncoplot}}
}
