% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncostrip.R
\name{oncostrip}
\alias{oncostrip}
\title{draw an oncostrip similar to cBioportal oncoprinter output.}
\usage{
oncostrip(maf, genes = NULL, sort = TRUE, sortByAnnotation = FALSE,
  annotation = NULL, annotationColor = NULL, removeNonMutated = FALSE,
  top = 5, showTumorSampleBarcodes = FALSE, colors = NULL)
}
\arguments{
\item{maf}{an \code{\link{MAF}} object generated by \code{read.maf}}

\item{genes}{draw oncoprint for these genes. default NULL. Plots top 5 genes.}

\item{sort}{logical sort oncomatrix for enhanced visualization. Defaults to TRUE.}

\item{sortByAnnotation}{logical sort oncomatrix by provided annotations. Defaults to FALSE. This is mutually exclusive with \code{sort}.}

\item{annotation}{\code{data.frame} with first column containing Tumor_Sample_Barcodes and rest of columns with annotations.}

\item{annotationColor}{list of colors to use for annotation. Default NULL.}

\item{removeNonMutated}{Logical. If \code{TRUE} removes samples with no mutations in the oncoplot for better visualization.}

\item{top}{how many top genes to be drawn. defaults to 5.}

\item{showTumorSampleBarcodes}{logical to include sample names.}

\item{colors}{named vector of colors for each Variant_Classification.}
}
\value{
None.
}
\description{
draw an oncostrip similar to cBioportal oncoprinter output.
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf, removeSilent = TRUE, useAll = FALSE)
dev.new()
oncostrip(maf = laml, genes = c('NPM1', 'RUNX1'), removeNonMutated = TRUE)

}
\seealso{
\code{\link{oncoplot}}
}

