% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cbs_segs.R
\name{plotCBSsegments}
\alias{plotCBSsegments}
\title{Plots segmented copy number data.}
\usage{
plotCBSsegments(cbsFile = NULL, maf = NULL, tsb = NULL, chr = NULL,
  savePlot = FALSE, width = 6, height = 3, labelAll = FALSE,
  genes = NULL, ref.build = "hg19", writeTable = FALSE,
  removeXY = FALSE, color = NULL)
}
\arguments{
\item{cbsFile}{CBS segmented copy number file. Column names should be Sample, Chromosome, Start, End, Num_Probes and Segment_Mean (log2 scale).}

\item{maf}{optional \code{\link{MAF}}}

\item{tsb}{If segmentation file contains many samples (as in gistic input), specify sample name here. Defualt plots all samples. If you are maping maf, make sure sample names in
Sample column of segmentation file matches to those Tumor_Sample_Barcodes in MAF.}

\item{chr}{Just plot this chromosome.}

\item{savePlot}{If true plot is saved as pdf.}

\item{width}{width of plot}

\item{height}{height of plot}

\item{labelAll}{If true and if maf object is specified, maps all mutataions from maf onto segments. Default FALSE, maps only variants on copy number altered regions.}

\item{genes}{highlight only these variants}

\item{ref.build}{Reference build for chromosome sizes. Can be hg18, hg19 or hg38. Default hg19.}

\item{writeTable}{If true and if maf object is specified, writes plot data with each variant and its corresponding copynumber to an output file.}

\item{removeXY}{don not plot sex chromosomes.}

\item{color}{Manually specify color scheme for chromosomes. Default NULL.}
}
\value{
ggplot object
}
\description{
Plots segmented copy number data.
}
\details{
this function takes segmented copy number data and plots it. If MAF object is specified, all mutations are highlighted on the plot.
}
\examples{
tcga.ab.009.seg <- system.file("extdata", "TCGA.AB.3009.hg19.seg.txt", package = "maftools")
plotCBSsegments(cbsFile = tcga.ab.009.seg)

}
