% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGisticResults.R
\name{plotGisticResults}
\alias{plotGisticResults}
\title{Plot gistic results as a bubble plot or as a genomic segment plot.}
\usage{
plotGisticResults(gistic = NULL, gis.scores = NULL, fdrCutOff = 0.1,
  markBands = NULL, color = NULL, ref.build = "hg19",
  cytobandOffset = 0.03, file = NULL, width = 6, height = 5,
  txtSize = 3)
}
\arguments{
\item{gistic}{an object of class \code{GISTIC} generated by \code{readGistic}}

\item{gis.scores}{optional scores.gistic file generated by GISTIC. If provided plot would be a histogram of amplifications and deletions along the genome. Default NULL}

\item{fdrCutOff}{fdr cutoff to use. Default 0.1}

\item{markBands}{any cytobands to label.}

\item{color}{colors for Amp and Del events.}

\item{ref.build}{reference build. Could be hg18, hg19 or hg38.}

\item{cytobandOffset}{if scores.gistic file is given use this to adjust cytoband size.}

\item{file}{if given saves plot as a pdf.}

\item{width}{width of the file to be saved.}

\item{height}{height of the file to be saved.}

\item{txtSize}{label size for bubbles.}
}
\value{
nothing
}
\description{
Plots significantly altered cytobands as a function of number samples in which it is altered and number genes it contains. Size of each bubble is according to -log10 transformed q values.
Optionally if scores.gistic file is provided, a genomic plot is generated with segments highlighting signififcant Amplifications and Deletion regions.
}
\examples{
all.lesions <- system.file("extdata", "all_lesions.conf_99.txt", package = "maftools")
amp.genes <- system.file("extdata", "amp_genes.conf_99.txt", package = "maftools")
del.genes <- system.file("extdata", "del_genes.conf_99.txt", package = "maftools")
laml.gistic = readGistic(gisticAllLesionsFile = all.lesions, gisticAmpGenesFile = amp.genes, gisticDelGenesFile = del.genes, isTCGA = TRUE)
plotGisticResults(gistic = laml.gistic)
}
