% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMafSummary.R
\name{plotmafSummary}
\alias{plotmafSummary}
\title{Plots maf summary.}
\usage{
plotmafSummary(maf, file = NULL, width = 6, height = 5,
  rmOutlier = TRUE, addStat = NULL, rmSilent = TRUE,
  showBarcodes = FALSE, textSize = 2, color = NULL)
}
\arguments{
\item{maf}{an MAF object generated by \code{\link{read.maf}}}

\item{file}{If given pdf file will be generated.}

\item{width}{plot parameter for output file.}

\item{height}{plot parameter for output file.}

\item{rmOutlier}{If TRUE removes outlier from boxplot.}

\item{addStat}{Can be either mean or median. Default NULL.}

\item{rmSilent}{remove silent variants from the plot.}

\item{showBarcodes}{include sample names in the top bar plot.}

\item{named}{vector of colors for each Variant_Classification.}
}
\description{
Plots maf summary.
}
\examples{
plotmafSummary(maf = laml, rmSilent = T, addStat = 'median')
}

