% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMafSummary.R
\name{plotmafSummary}
\alias{plotmafSummary}
\title{Plots maf summary.}
\usage{
plotmafSummary(maf, file = NULL, rmOutlier = TRUE, width = 6,
  height = 5, addStat = NULL, showBarcodes = FALSE, textSize = 2,
  color = NULL)
}
\arguments{
\item{maf}{an \code{\link{MAF}} object generated by \code{\link{read.maf}}}

\item{file}{If given pdf file will be generated.}

\item{rmOutlier}{If TRUE removes outlier from boxplot.}

\item{width}{plot parameter for output file.}

\item{height}{plot parameter for output file.}

\item{addStat}{Can be either mean or median. Default NULL.}

\item{showBarcodes}{include sample names in the top bar plot.}

\item{textSize}{font size if showBarcodes is TRUE. Default 2.}

\item{color}{named vector of colors for each Variant_Classification.}
}
\value{
Prints plot.
}
\description{
Plots maf summary.
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf, useAll = FALSE)
plotmafSummary(maf = laml, addStat = 'median')
}
\seealso{
\code{\link{read.maf}} \code{\link{MAF}}
}

