% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignatures.R
\name{plotSignatures}
\alias{plotSignatures}
\title{Plots decomposed mutational signatures.}
\usage{
plotSignatures(nmfRes = NULL, contributions = FALSE, ...)
}
\arguments{
\item{nmfRes}{results from \code{\link{extractSignatures}}}

\item{contributions}{If TRUE plots contribution of signatures in each sample.}

\item{...}{further plot options passed to \code{\link{barplot2}}}
}
\value{
ggplot object if contributions is TRUE
}
\description{
Plots decomposed mutational signatures as a barplot.
}

