% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignatures.R
\name{plotSignatures}
\alias{plotSignatures}
\title{Plots decomposed mutational signatures.}
\usage{
plotSignatures(nmfRes = NULL)
}
\arguments{
\item{nmfRes}{results from \code{\link{extractSignatures}}}
}
\value{
None.
}
\description{
Plots decomposed mutational signatures as a barplot.
}

