% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignatures.R
\name{plotSignatures}
\alias{plotSignatures}
\title{Plots decomposed mutational signatures.}
\usage{
plotSignatures(nmfRes = NULL, contributions = FALSE, color = NULL, ...)
}
\arguments{
\item{nmfRes}{results from \code{\link{extractSignatures}}}

\item{contributions}{If TRUE plots contribution of signatures in each sample.}

\item{color}{colors for each Ti/Tv conversion class. Default NULL}

\item{...}{further plot options passed to \code{\link{barplot}}}
}
\value{
ggplot object if contributions is TRUE
}
\description{
Plots decomposed mutational signatures as a barplot.
}
\seealso{
\code{\link{trinucleotideMatrix}}
}

