% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignatures.R
\name{plotSignatures}
\alias{plotSignatures}
\title{Plots decomposed mutational signatures}
\usage{
plotSignatures(nmfRes = NULL, contributions = FALSE, color = NULL,
  patient_order = NULL, title_size = 1.2, ...)
}
\arguments{
\item{nmfRes}{results from \code{\link{extractSignatures}}}

\item{contributions}{If TRUE plots contribution of signatures in each sample.}

\item{color}{colors for each Ti/Tv conversion class. Default NULL}

\item{patient_order}{User defined ordering of samples. Default NULL.}

\item{title_size}{size of title. Default 1.3}

\item{...}{further plot options passed to \code{\link{barplot}}}
}
\value{
ggplot object if contributions is TRUE
}
\description{
Takes results from \code{\link{extractSignatures}} and plots decomposed mutational signatures as a barplot.
}
\seealso{
\code{\link{trinucleotideMatrix}} \code{\link{plotSignatures}}
}
