% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTiTv.R
\name{plotTiTv}
\alias{plotTiTv}
\title{Plot Transition and Trasnversion ratios.}
\usage{
plotTiTv(res = NULL, file = NULL, width = 6, height = 5)
}
\arguments{
\item{res}{results generated by \code{\link{titv}}}

\item{file}{basename for output file name. If given pdf will be generated.}

\item{width}{width of the plot, in inches.}

\item{height}{height of the plot, in inches.}
}
\description{
Takes results generated from \code{titv} and plots the Ti/Tv ratios and contributions of 6 mutational conversion classes in each sample.
}

