% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vaf.R
\name{plotVaf}
\alias{plotVaf}
\title{Plots vaf distribution of genes}
\usage{
plotVaf(maf, vafCol = NULL, genes = NULL, violin = FALSE, top = 10,
  orderByMedian = TRUE, flip = FALSE, fn = NULL, width = 6,
  height = 5)
}
\arguments{
\item{maf}{an \code{\link{MAF}} object generated by \code{\link{read.maf}}}

\item{vafCol}{manually specify column name for vafs. Default looks for column 't_vaf'}

\item{genes}{specify genes for which plots has to be generated}

\item{violin}{if TRUE plots violin plot}

\item{top}{if \code{genes} is NULL plots top n number of genes. Defaults to 5.}

\item{orderByMedian}{Orders genes by decreasing median VAF. Default TRUE}

\item{flip}{if TRUE, flips axes. Default FALSE}

\item{fn}{Filename. If given saves plot as a output pdf. Default NULL.}

\item{width}{Width of plot to be saved. Default 6}

\item{height}{Height of plot to be saved. Default 5}
}
\value{
ggplot object which can be further modified.
}
\description{
Plots vaf distribution of genes as a boxplot or violinplot.
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf, removeSilent = TRUE, useAll = FALSE)
plotVaf(maf = laml, vafCol = 'i_TumorVAF_WU')

}
