% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepAscat.R
\name{prep_ascat}
\alias{prep_ascat}
\title{Prepare input files for ASCAT}
\usage{
prep_ascat(
  t_counts = NULL,
  n_counts = NULL,
  sample_name = NA,
  min_depth = 15,
  normalize = FALSE
)
}
\arguments{
\item{t_counts}{read counts from tumor generated by \code{\link{gtSNPs}}}

\item{n_counts}{read counts from normal generated by \code{\link{gtSNPs}}}

\item{sample_name}{Sample name. Used as a basename for output files. Default `NA`, parses from `t_counts` file.}

\item{min_depth}{Min read depth required to consider a marker. Default 15}

\item{normalize}{If TRUE, normalizes for library size. Default FALSE}
}
\value{
An \code{\link{ascat.loadData}} object; ascat data structure
}
\description{
Function takes the output from \code{\link{gtSNPs}} and generates `logR` and `BAF` files required for ASCAT analysis.
}
\details{
The function will filter SNPs with low coverage (default <15), estimate BAF, logR, and generates the input files for ASCAT. In addition, it will run `ASCAT::ascat.loadData()` and `ASCAT::ascat.plotRawData()` for you and returns the `ASCAT` object that can be further processed with ASCAT functions. Alternatively, logR file can be segmented with \code{\link{segment_logR}}
}
\seealso{
\code{\link{gtSNPs}} \code{\link{prep_ascat_t}} \code{\link{segment_logR}}
}
