% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maf_dt.R
\name{read.maf}
\alias{read.maf}
\title{Read MAF files.}
\usage{
read.maf(maf, removeSilent = TRUE, useAll = FALSE)
}
\arguments{
\item{maf}{tab delimited MAF file. File can also be gz compressed.}

\item{removeSilent}{logical. Whether to discard silent (with no functional impact) mutations ("Silent","Intron","RNA","3'UTR"). Default is TRUE.}

\item{useAll}{logical. Whether to use all variants irrespective of values in Mutation_Status. Defaults to False. Only uses with values Somatic.}
}
\value{
An object of class MAF.
}
\description{
Takes tab delimited MAF (can be plain text or gz compressed) file as an input and summarizes it in various ways. Also creates oncomatrix - helpful for visualization.
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf, removeSilent = TRUE, useAll = FALSE)
}

