% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgacompare.R
\name{tcgaCompare}
\alias{tcgaCompare}
\title{Compare mutation load against TCGA cohorts}
\usage{
tcgaCompare(
  maf,
  capture_size = NULL,
  tcga_capture_size = 50,
  cohortName = NULL,
  tcga_cohorts = NULL,
  primarySite = FALSE,
  col = c("gray70", "black"),
  bg_col = c("#EDF8B1", "#2C7FB8"),
  medianCol = "red",
  decreasing = FALSE,
  logscale = TRUE,
  rm_hyper = FALSE
)
}
\arguments{
\item{maf}{\code{\link{MAF}} object(s) generated by \code{\link{read.maf}}}

\item{capture_size}{capture size for input MAF in MBs. Default NULL. If provided plot will be scaled to mutations per mb. TCGA capture size is assumed to be 50mb.}

\item{tcga_capture_size}{capture size for TCGA cohort in MB. Default 50}

\item{cohortName}{name for the input MAF cohort. Default "Input"}

\item{tcga_cohorts}{restrict tcga data to these cohorts.}

\item{primarySite}{If TRUE uses primary site of cancer as labels instead of TCGA project IDs. Default FALSE.}

\item{col}{color vector for length 2 TCGA cohorts and input MAF cohort. Default gray70 and black.}

\item{bg_col}{background color. Default'#EDF8B1', '#2C7FB8'}

\item{medianCol}{color for median line. Default red.}

\item{decreasing}{Default FALSE. Cohorts are arranged in increasing mutation burden.}

\item{logscale}{Default TRUE}

\item{rm_hyper}{Remove hyper mutated samples (outliers)? Default FALSE}
}
\value{
data.table with median mutations per cohort
}
\description{
Compares mutation load in input MAF against all of 33 TCGA cohorts derived from MC3 project.
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
tcgaCompare(maf = laml, cohortName = "AML")
}
