% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driverBP.R
\name{tcgaDriverBP}
\alias{tcgaDriverBP}
\title{Compare genes to known TCGA drivers and their biological pathways}
\usage{
tcgaDriverBP(m, genes = NULL, top = 20, fontSize = 0.7)
}
\arguments{
\item{m}{an \code{\link{MAF}} object}

\item{genes}{genes to compare. Default `NULL`.}

\item{top}{Top number of genes to use. Mutually exclusive with `genes` argument. Default 20}

\item{fontSize}{Default 0.7}
}
\description{
A small function which uses known cancer driver genes and their associatd pathways from TCGA cohorts. See reference for details
}
\examples{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read.maf(maf = laml.maf)
tcgaDriverBP(m = laml)
}
\references{
Bailey MH, Tokheim C, Porta-Pardo E, et al. Comprehensive Characterization of Cancer Driver Genes and Mutations . Cell. 2018;173(2):371–385.e18. doi:10.1016/j.cell.2018.02.060
}
