% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titv.R
\name{titv}
\alias{titv}
\title{Classifies SNPs into transitions and transversions}
\usage{
titv(maf, plot = TRUE)
}
\arguments{
\item{maf}{output generated by \code{read.maf}}

\item{plot}{plots a titv fractions. default TRUE.}
}
\value{
list of \code{data.frame}s with Transitions and Transversions summary.
}
\description{
takes output generated by read.maf and classifies Single Nucleotide Variants into Transitions and Transversions.
}

