% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr.R
\name{vcr}
\alias{vcr}
\title{Samll internal function to make complex events.}
\usage{
vcr(xstr, gis = FALSE)
}
\arguments{
\item{xstr}{charcter to split}

\item{gis}{Is input from gistic. Logical.}
}
\value{
split string
}
\description{
Samll internal function to make complex events. Ignore this.
}

